+++ models/kanji.rb
@@ -1,0 +1,5 @@
+ActiveRecord::Base.configurations = YAML.load_file('database.yml')
+ActiveRecord::Base.establish_connection(:development)
+
+class Kanji < ActiveRecord::Base
+end
