+++ views/management.erb
@@ -3,7 +3,7 @@
     <th>id</th>
     <th>name</th>
     <th>email</th>
-    <th>password</th>
+    <th>password_digest</th>
   </tr>
 
   <% @users.each do |user| %>
